pyproject.toml
[project]
name = "promptgres"
version = "0.1.0"
description = "A personal toolkit for rapid SQL query generation from stakeholder requirements using AI assistance."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
